+++ src/modules/liquidity/pages/v3-item-page/helpers/map-position.ts
@@ -1,22 +1,10 @@
 import BigNumber from 'bignumber.js';
-import cx from 'classnames';
 
-import { AppRootRoutes } from '@app.router';
 import { IS_NETWORK_MAINNET } from '@config/config';
 import { TESTNET_EXCHANGE_RATE, ZERO_AMOUNT_BN } from '@config/constants';
-import { LiquidityRoutes, LiquiditySubroutes } from '@modules/liquidity/liquidity-routes.enum';
 import { LiquidityV3Position } from '@modules/liquidity/types';
-import {
-  getSumOfNumbers,
-  getTokenDecimals,
-  getTokensNames,
-  isExist,
-  multipliedIfPossible,
-  toAtomic,
-  toReal
-} from '@shared/helpers';
-import { ActiveStatus, Optional, Token } from '@shared/types';
-import { i18n } from '@translation';
+import { getSumOfNumbers, getTokenDecimals, isExist, multipliedIfPossible, toAtomic, toReal } from '@shared/helpers';
+import { Optional, Token } from '@shared/types';
 
 import { convertToAtomicPrice } from './convert-to-atomic-price';
 
@@ -25,24 +13,16 @@ const MOCK_NON_ZERO_TOKEN_Y_REAL_DEPOSIT = new BigNumber('0.3');
 const MOCK_TOKEN_X_REAL_FEES = new BigNumber('0.01');
 const MOCK_TOKEN_Y_REAL_FEES = new BigNumber('0.02');
 
-interface RangeLabelClasses {
-  className: string;
-  inRangeClassName: string;
-}
-
 export const mapPosition = (
   tokenX: Token,
   tokenY: Token,
   currentRealPrice: Optional<BigNumber>,
   getTokenExchangeRate: (token: Token) => Optional<BigNumber>,
-  poolId: Optional<string>,
-  rangeLabelClasses: RangeLabelClasses
+  poolId: Optional<string>
 ) => {
-  const { className: rangeLabelClassName, inRangeClassName } = rangeLabelClasses;
   const tokenXExchangeRate = IS_NETWORK_MAINNET ? getTokenExchangeRate(tokenX) : TESTNET_EXCHANGE_RATE;
   const tokenYExchangeRate = IS_NETWORK_MAINNET ? getTokenExchangeRate(tokenY) : TESTNET_EXCHANGE_RATE;
   const tokenPriceDecimals = getTokenDecimals(tokenY) - getTokenDecimals(tokenX);
-  const tokensNames = getTokensNames([tokenY, tokenX]);
 
   return (position: LiquidityV3Position) => {
     const minRange = toReal(convertToAtomicPrice(position.lower_tick.sqrt_price), tokenPriceDecimals);
@@ -73,73 +53,19 @@ export const mapPosition = (
     const isInRange = isExist(currentRealPrice) && currentRealPrice.gte(minRange) && currentRealPrice.lte(maxRange);
 
     return {
-      href: `${AppRootRoutes.Liquidity}${LiquidityRoutes.v3}/${poolId}/${
-        LiquiditySubroutes.positions
-      }/${position.id.toFixed()}`,
-      inputToken: [tokenX, tokenY],
-      status: null,
-      isNew: false,
-      labels: [
-        {
-          contentClassName: cx(rangeLabelClassName, isInRange && inRangeClassName),
-          status: ActiveStatus.ACTIVE,
-          label: isInRange ? i18n.t('liquidity|inRange') : i18n.t('liquidity|notActive')
-        }
-      ],
-      itemStats: [
-        {
-          cellName: i18n.t('liquidity|minPrice'),
-          amounts: {
-            amount: minRange,
-            dollarEquivalent: null,
-            currency: tokensNames
-          },
-          DTI: 'minPrice',
-          tooltip: i18n.t('liquidity|minPriceTooltip')
-        },
-        {
-          cellName: i18n.t('liquidity|maxPrice'),
-          amounts: {
-            amount: maxRange,
-            dollarEquivalent: null,
-            currency: tokensNames
-          },
-          DTI: 'maxPrice',
-          tooltip: i18n.t('liquidity|maxPriceTooltip')
-        },
-        {
-          cellName: i18n.t('liquidity|Deposit'),
-          amounts: {
-            amount: depositUsd,
-            dollarEquivalent: depositUsd,
-            currency: tokensNames,
-            dollarEquivalentOnly: true
-          },
-          DTI: 'deposit',
-          tooltip: i18n.t('liquidity|depositTooltip')
-        },
-        {
-          cellName: i18n.t('liquidity|collectedFees'),
-          amounts: {
-            amount: collectedFeesUsd,
-            dollarEquivalent: collectedFeesUsd,
-            currency: tokensNames,
-            dollarEquivalentOnly: true
-          },
-          DTI: 'collectedFees',
-          tooltip: i18n.t('liquidity|collectedFeesTooltip')
-        }
-      ],
-      itemDTI: `liquidity-item-v3-${position.id.toFixed()}`,
       collectedFeesUsd,
       depositUsd,
       minRange,
       maxRange,
       isInRange,
+      tokenX,
+      tokenY,
       tokenXDeposit,
       tokenYDeposit,
       tokenXFees,
-      tokenYFees
+      tokenYFees,
+      poolId,
+      id: position.id
     };
   };
 };
